--- desktopapp/output/data-out.yaml
@@ -1,102 +1,0 @@
-General:
-  ListenAddress: 127.0.0.1
-  ListenPort: 999
-  TLS:
-    Enabled: false
-    PrivateKey: tls/server.key
-    Certificate: 9996
-    RootCAs:
-      - tls/ca.crt
-    ClientAuthRequired: false
-    ClientRootCAs: null
-  Keepalive:
-    ServerMinInterval: 60s
-    ServerInterval: 7200s
-    ServerTimeout: 20s
-  Cluster:
-    SendBufferSize: 10
-    ClientCertificate: null
-    ClientPrivateKey: null
-    ListenPort: null
-    ListenAddress: null
-    ServerCertificate: null
-    ServerPrivateKey: null
-  BootstrapMethod: file
-  BootstrapFile: null
-  LocalMSPDir: msp
-  LocalMSPID: SampleOrg
-  Profile:
-    Enabled: false
-    Address: '0.0.0.0:6060'
-  BCCSP:
-    Default: SW
-    SW:
-      Hash: SHA2
-      Security: 256
-      FileKeyStore: null
-      KeyStore: null
-    PKCS11:
-      Library: null
-      Label: null
-      Pin: null
-      Hash: null
-      Security: null
-      FileKeyStore: null
-      KeyStore: null
-  Authentication:
-    TimeWindow: 15m
-FileLedger:
-  Location: /var/hyperledger/production/orderer
-  Prefix: hyperledger-fabric-ordererledger
-Kafka:
-  Retry:
-    ShortInterval: 5s
-    ShortTotal: 10m
-    LongInterval: 5m
-    LongTotal: 12h
-    NetworkTimeouts:
-      DialTimeout: 10s
-      ReadTimeout: 10s
-      WriteTimeout: 10s
-    Metadata:
-      RetryBackoff: 250ms
-      RetryMax: 3
-    Producer:
-      RetryBackoff: 100ms
-      RetryMax: 3
-    Consumer:
-      RetryBackoff: 2s
-  Topic:
-    ReplicationFactor: 3
-  Verbose: false
-  TLS:
-    Enabled: false
-    PrivateKey: null
-    Certificate: null
-    RootCAs: null
-  SASLPlain:
-    Enabled: false
-    User: null
-    Password: null
-  Version: null
-Debug:
-  BroadcastTraceDir: null
-  DeliverTraceDir: null
-Operations:
-  ListenAddress: '127.0.0.1:8443'
-  TLS:
-    Enabled: false
-    Certificate: null
-    PrivateKey: null
-    ClientAuthRequired: false
-    ClientRootCAs: null
-Metrics:
-  Provider: disabled
-  Statsd:
-    Network: udp
-    Address: '127.0.0.1:8125'
-    WriteInterval: 30s
-    Prefix: null
-Consensus:
-  WALDir: /var/hyperledger/production/orderer/etcdraft/wal
-  SnapDir: /var/hyperledger/production/orderer/etcdraft/snapshot
