--- desktopapp/output/user-input.json
@@ -1,1 +1,0 @@
-{"port":999,"Certificate":9996}
