+++ desktopapp/src/lib/ca-data.d.ts
@@ -1,0 +1,6 @@
+// Interface for call data class
+declare module 'ConfigData' {
+    interface ConfigData {
+        getComment: () => string;
+    }
+}
