--- desktopapp/tests/README.md
@@ -1,1 +1,0 @@
-this file make for git can see folder
